requirements.txt
requests
streamlit
pandas
python-dotenv
